+++ 2 Search and Sort Algorithms/CSC395 Module3 Queue/.vs/CSC395 Module3/xs/UserPrefs.xml
@@ -1,19 +1,25 @@
 ﻿<Properties StartupConfiguration="{D5416C46-3F7D-47AE-B5B1-21764F48F23F}|Default">
-  <MonoDevelop.Ide.Workbench ActiveDocument="CSC395 Module3/Program.cs">
+  <MonoDevelop.Ide.Workbench ActiveDocument="CSC395 Module3/Stack.cs">
     <Files>
       <File FileName="CSC395 Module3/Program.cs" Line="1" Column="1" />
+      <File FileName="CSC395 Module3/Stack.cs" Line="1" Column="1" />
+      <File FileName="CSC395 Module3/Node.cs" Line="1" Column="1" />
+      <File FileName="CSC395 Module3/SinglyLinkedList.cs" Line="1" Column="1" />
+      <File FileName="CSC395 Module3 Queue Status" />
     </Files>
     <Pads>
       <Pad Id="ProjectPad">
         <State name="__root__">
           <Node name="CSC395 Module3" expanded="True">
-            <Node name="CSC395 Module3" expanded="True" selected="True" />
+            <Node name="CSC395 Module3" expanded="True">
+              <Node name="Stack.cs" selected="True" />
+            </Node>
           </Node>
         </State>
       </Pad>
       <Pad Id="ClassPad">
         <State name="__root__">
-          <Node name="CSC395 Module3" expanded="True" selected="True" />
+          <Node name="CSC395 Module3" selected="True" />
         </State>
       </Pad>
     </Pads>
