+++ 2 Search and Sort Algorithms/CSC395 Module3 Queue/CSC395 Module3/Node.cs
@@ -1,0 +1,20 @@
+﻿using System;
+namespace CSC395_Module3
+{
+    class Node
+    {
+        //data
+        public int value { get; set; }
+        public Node next;
+
+        //methods
+
+        //ctor
+        public Node(int value)
+        {
+            this.value = value;
+            this.next = null;
+        }
+
+    }
+}
