+++ 2 Search and Sort Algorithms/CSC395 Module3 Queue/CSC395 Module3/Queue2.cs
@@ -1,0 +1,10 @@
+﻿using System;
+namespace CSC395_Module3
+{
+    public class Queue2
+    {
+        public Queue2()
+        {
+        }
+    }
+}
