+++ 2 Search and Sort Algorithms/CSC395 Module3 Queue/CSC395 Module3/Stack.cs
@@ -1,0 +1,39 @@
+﻿using System;
+namespace CSC395_Module3
+{
+    public class Stack
+    {
+        //data - we'll base our stack on a linked list
+        SinglyLinkedList myStackData;
+
+        //
+        public bool IsEmpty()
+        {
+            return myStackData.IsEmpty();
+        }
+
+        public void Push(int val)
+        {
+            myStackData.AddFirst(val);
+        }
+
+        public void Pop()
+        {
+            myStackData.RemoveFirst();
+        }
+
+
+        public int Peek()
+        {
+            return myStackData.first.value;
+        }
+
+
+        //ctor  
+        public Stack()
+        {
+            myStackData = new SinglyLinkedList();
+        }
+
+    }
+}
